task#7882: Multiple improves to the csv2xml library (#1)
* [IMP] Be able to create openerp modules vesion 8.0 and 10.0

add manifest_file as one of the valid ones to identify a odoo module

* [ADD] new required argument named new_path that will set the path were the new
files will be created. Either xml (update subcomand) or csv (create subcomand)

Also remove the constraint about the required data folder. This is not required
anymore since we are using a new argument --new-path that set the path where
the new files will be added.

* [FIX] remove an orphan print inside the code.

* [ADD] remove part that updates the module descriptor with the new data.

This section has no sence with the new --new-path argument.
For that reason is comment until a new way to manage this became real.